requirements.txt
amqp==2.6.1
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
distlib==0.3.6
Django==3.2.13
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-redis==5.2.0
django-tables2==2.4.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
filelock==3.8.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iso8601==1.0.2
kombu==4.6.11
packaging==21.3
pipenv==2022.9.4
platformdirs==2.5.2
promise==2.3
prompt-toolkit==3.0.31
PyJWT==1.7.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.2
vine==1.3.0
virtualenv==20.16.4
virtualenv-clone==0.5.7
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.14.1